backend/requirements.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
configparser==3.5.0
Django==1.11.17
django-cors-headers==2.4.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
enum34==1.1.6
futures==3.2.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
passlib==1.7.1
PyJWT==1.7.1
pylint==1.9.3
pytz==2018.7
singledispatch==3.4.0.3
six==1.12.0
wrapt==1.10.11
